+++ app/src/main/java/com/naldana/ejemplo10/pojo/Coin.kt
@@ -3,6 +3,6 @@ package com.naldana.ejemplo10.pojo
 data class Coin(
     var name: String = "N/A",
     var country: String = "N/A",
-    var year: Long = 0,
+    var year: Int = 0,
     var available: Boolean = false
 )
